requirements.txt
pycryptodome>=3.9
roca-detect>=1.2
